add make to required VM packages and fix init recipe

+++ Makefile
@@ -14,10 +14,7 @@ install:
 # initialize the app with the session key
 init:
 	mkdir -p app-data
-	if [[ ! -d app-data || ! -f app-data/session-secret-key.key ]]
-	then
-		docker compose -f docker-compose.init.yaml run --rm --build create-session
-	fi
+	test ! -d app-data -o ! -f app-data/session-secret-key.key && docker compose -f docker-compose.init.yaml run --rm --build create-session
 
 
 # DEV only
